backend/requirements.txt
cryptography==41.0.7
pycryptodome==3.19.0
argon2-cffi==23.1.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
bcrypt==4.1.1
PyJWT==2.10.1
